+++ cli/aux/aux.h
@@ -1,0 +1,9 @@
+#ifndef TEC_AUX_H
+#define TEC_AUX_H
+
+#define PROGRAM     "tec"
+
+// TODO: Get rid of it (set options with default boolean values)
+#define NONEBOOL        -1      /* Not yet set boolean value */
+
+#endif